Reorganize for clarity
* Move all components specific to home page into (home) route group
* Move all common components into _common private folder and
  add ts alias to it for convenience
* Remove unused rocks stack svg
