+++ app/(home)/AboutSection.tsx
@@ -1,5 +1,5 @@
 import { type ReactNode } from 'react';
-import { cn } from '@/lib/cn';
+import { cn } from '@/cn';
 import { Icon } from '@/components/Icon';
 import { RemoteImage, type RemoteImageId } from '@/components/RemoteImage';
