+++ app/(home)/ContactLinks.tsx
@@ -1,5 +1,5 @@
 import { type ComponentProps } from 'react';
-import { cn } from '@/lib/cn';
+import { cn } from '@/cn';
 import { Icon } from '@/components/Icon';
 
 function LinksList({
