+++ app/_common/components/Badge.tsx
@@ -1,5 +1,5 @@
 import { type ReactNode } from 'react';
-import { cn } from '@/lib/cn';
+import { cn } from '@/cn';
 
 export function Badge({ children }: { children: ReactNode }) {
   return (
