+++ app/_common/components/Popover.tsx
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import * as PopoverPrimitive from '@radix-ui/react-popover';
 
-import { cn } from '@/lib/cn';
+import { cn } from '@/cn';
 
 const Popover = PopoverPrimitive.Root;
