+++ components.json
@@ -11,6 +11,6 @@
   },
   "aliases": {
     "components": "@/components",
-    "utils": "@/lib/cn"
+    "utils": "@/cn"
   }
 }
