+++ tsconfig.json
@@ -20,7 +20,8 @@
       }
     ],
     "paths": {
-      "@/*": ["./*"]
+      "@configs/*": ["./*"],
+      "@/*": ["./app/_common/*"]
     }
   },
   "include": [
